Allow alternative isnan/isnanf implementations
Many compilers have __builtin_* implementations of <math.h> functions
which can be used without depending on <math.h> (or libm).  This patch
provides a way for SIMDe to make use of them, as well as other
implementations (just define simde_isnan/simde_isnanf prior to
including SIMDe).

Initially, this helps increase the number of functions we can use in
non-hosted environments, such as clang's built-in WASM support.  In the
future, though, I'd like to add similar code for other libm functions
so implementations can provide their own implementations of certain
functions.  Those implementations may be faster approximations (like
those in Cephes and VDT).

I may end up moving this into a separate header as it grows (especially
with SVML support expanding), and possibly add a version of this to
portable-snippets.