Rewrite feature detection macros.
The new version is much simpler, largely because we're recreating less
in each file; most of the logic has been centralized in
simde-features.h.

The down side is that we do lose the ability to enable/disable specific
implementations on a per-extension basis.  For example, you can no
longer disable the NEON implementation of SSE2 but leave it enabled
for SSE.